Experience page ready
All information in the experience page is correct and accurate, the styling is also satisfactory now (with room for improvement pending consultation with a designer). There are no images on this page because I couldn't figure out where they would go, so for now it will be a text only page.