Created Data.json
Data.json file in Assets folder now holds list of programming skills. The Home and About page now read from this file to display the programming skills. This means that in the future if I want to add a new skill, I only need to add it to the Data.json file and it will update everywhere.