Display single project
Display a single project on the projects page. Not much else to say, still needs a lot of work on styling the projects, i want to add an overlay to them at some point. The projects page as a whole still has quite a bit of work left, such as a skill filtering feature