AP-2: Selected Navbar Button is green
the selected navbar button now has green text when not hovered over. This change will always illustrate to the user where they are in the website. Additionally, when the navbar buttons are hovered over, the size of the button slightly increases and the text becomes bold, which gives a very animated look to the navbar. Overall makes the site much more polished.