AP-49: Project Details Overhaul
The project details page now uses a unique URL id parameter to display the correct project instead of using local storage. This is a much better implementation because it allows every project to have their own unique project details page making it possible to share links to a specific project.