AP-52: Project Filters don't deselect on mobile
On mobile version, the selected project filter doesn't deselect when the user performs a browser back to a previous selected filter. I identified the issue as the the hover attribute of the filters, due to the way mobile browsers work the currently selected filter remains actively focused even during a browser back. I removed the hover functionality of the project filters on smaller screens. Unfortunately this problem will presist on any large screen touch device.